refactor: split activate() from handlers
